Factor out AsyncRewriter
